Re-aranged components call
